+++ crates/tool-lookup/changelog.md
@@ -1,2 +1,7 @@
+# 1.0.1 (2025-10-03)
+- Added support for matching by exact file name for dotfiles or bare names.
+  - Pattern ".env" matches basename ".env"
+  - Pattern "env" matches basename "env"
+
 # 1.0.0 (2025-10-03)
 - Initial release.